backend/requirements.dev.txt
anyio
Django
django-cors-headers
django-extensions
django-filter
django-js-asset
django-redis
django-timezone-field
djangorestframework
drf-standardized-errors
pillow
psutil
psycopg
psycopg2-binary
PyJWT
pyparsing
python-dateutil
python-dotenv
PyYAML
redis
regex
requests
six
gunicorn
django-debug-toolbar
django-notifications-hq~=1.8.3
django-jazzmin
django-environ
django-crispy-forms
crispy-bootstrap4
qrcode
social-auth-app-django